transaction: Add a default finalizer checking if transaction has ended
Add a dynamic check on transaction during its finalization, ensuring
that end is called and their resources are released.

It's a programmer error not to do that, so we should panic in such case.

Testing of this code-path requires wrapping a bit panic() because we
we can't recover from a panic happening in a goroutine.